+++ FA2sp/FA2sp.cpp
@@ -51,6 +51,8 @@ int ExtConfigs::RecentFileLimit;
 int ExtConfigs::MultiSelectionColor;
 bool ExtConfigs::RandomTerrainObjects;
 int ExtConfigs::MaxVoxelFacing;
+bool ExtConfigs::DDrawInVideoMem = true;
+bool ExtConfigs::DDrawEmulation = false;
 
 MultimapHelper Variables::Rules = { &CINI::Rules(), &CINI::CurrentDocument() };
 MultimapHelper Variables::FAData = { &CINI::FAData() };
@@ -137,6 +139,9 @@ void FA2sp::ExtConfigsInitialize()
 
 	// Disable it for now
 	ExtConfigs::MaxVoxelFacing = 8;
+
+	ExtConfigs::DDrawInVideoMem = CINI::FAData->GetBool("ExtConfigs", "DDrawInVideoMem");
+	ExtConfigs::DDrawEmulation = CINI::FAData->GetBool("ExtConfigs", "DDrawEmulation");
 }
 
 // DllMain