+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## RELEASE 1.5.0 (2022-12-XX)
 - Implement Auto Property Brush
+- New ***ExtConfig*** : `DDrawInVideoMem` = **BOOLEAN**, defaults to true
+- New ***ExtConfig*** : `DDrawEmulation` = **BOOLEAN**, defaults to false
 
 ## RELEASE 1.4.2 (2022-10-27)
 - Split `CncVxlRenderText`